+++ packages/user/src/model/invitations/resolver.ts
@@ -1,10 +1,7 @@
 import { formatDate } from "@dzangolab/fastify-slonik";
 import mercurius from "mercurius";
 import { createNewSession } from "supertokens-node/recipe/session";
-import {
-  emailPasswordSignUp,
-  getUsersByEmail,
-} from "supertokens-node/recipe/thirdpartyemailpassword";
+import { emailPasswordSignUp } from "supertokens-node/recipe/thirdpartyemailpassword";
 
 import { ROLE_ADMIN } from "../../constants";
 import computeInvitationExpiresAt from "../../lib/computeInvitationExpiresAt";
@@ -155,10 +152,18 @@ const Mutation = {
         return mercuriusError;
       }
 
-      // check if user of the email already exists
-      const emailUser = await getUsersByEmail(email);
+      const service = getInvitationService(config, database, dbSchema);
 
-      if (emailUser.length > 0) {
+      const emailFilter = {
+        key: "email",
+        operator: "eq",
+        value: email,
+      } as FilterInput;
+
+      const userCount = await service.count(emailFilter);
+
+      // check if user of the email already exists
+      if (userCount > 0) {
         const mercuriusError = new mercurius.ErrorWithProps(
           `User with email ${email} already exists`
         );
@@ -176,11 +181,11 @@ const Mutation = {
       const app = config.apps?.find((app) => app.id == appId);
 
       if (app) {
-        if (app.supportedRoles.includes(role)) {
+        if (app.supportedRoles.includes(invitationCreateInput.role)) {
           invitationCreateInput.appId = appId;
         } else {
           const mercuriusError = new mercurius.ErrorWithProps(
-            `App ${app.name} does not support role ${role}`
+            `App ${app.name} does not support role ${invitationCreateInput.role}`
           );
 
           return mercuriusError;
@@ -190,8 +195,6 @@ const Mutation = {
       if (Object.keys(payload || {}).length > 0) {
         invitationCreateInput.payload = JSON.stringify(payload);
       }
-
-      const service = getInvitationService(config, database, dbSchema);
 
       let invitation: Invitation | undefined;