+++ packages/user/src/model/users/handlers/disable.ts
@@ -1,7 +1,3 @@
-import { Error as STError } from "supertokens-node/recipe/session";
-import UserRoles from "supertokens-node/recipe/userroles";
-
-import { ROLE_ADMIN } from "../../../constants";
 import getUserService from "../../../lib/getUserService";
 
 import type { FastifyReply } from "fastify";
@@ -16,24 +12,6 @@ const disable = async (request: SessionRequest, reply: FastifyReply) => {
 
       return await reply.send({
         message: "you cannot disable yourself",
-      });
-    }
-
-    const roles = await request.session.getClaimValue(UserRoles.UserRoleClaim);
-
-    if (roles === undefined || !roles.includes(ROLE_ADMIN)) {
-      throw new STError({
-        type: "INVALID_CLAIMS",
-        message: "User is not an admin",
-        payload: [
-          {
-            id: UserRoles.UserRoleClaim.key,
-            reason: {
-              message: "wrong value",
-              expectedToInclude: ROLE_ADMIN,
-            },
-          },
-        ],
       });
     }